requirements.txt
absl-py==2.1.0
addict==2.4.0
appdirs==1.4.4
asttokens==2.4.0
astunparse==1.6.3
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
chumpy==0.70
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.1.1
cycler==0.11.0
Cython==3.0.2
debugpy==1.8.0
decorator==5.1.1
docker-pycreds==0.4.0
dominate==2.8.0
easydict==1.7
einops==0.7.0
exceptiongroup==1.1.3
executing==1.2.0
ffmpeg==1.4
filelock==3.9.0
flatbuffers==24.3.25
fonttools==4.42.1
fsspec==2023.12.2
gast==0.5.4
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.37
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
huggingface-hub==0.20.1
idna==3.4
imageio==2.31.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.26.0
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json-tricks==3.17.3
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.0
jupyter_core==5.5.0
keras==3.1.1
kiwisolver==1.4.5
lazy_loader==0.3
libclang==18.1.1
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mat4py==0.6.0
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
mmdet==3.2.0
mmengine==0.10.2
mmpose==1.3.0
mmpretrain==1.2.0
model-index==0.1.11
modelindex==0.0.2
mpmath==1.3.0
munch==4.0.0
munkres==1.1.4
namex==0.0.7
nest-asyncio==1.5.8
networkx==3.0
numpy==1.24.1
onnx==1.15.0
opencv-python==4.8.0.76
opt-einsum==3.3.0
optree==0.11.0
ordered-set==4.1.0
packaging==23.1
pandas==2.1.1
parso==0.8.3
pathtools==0.1.2
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==4.0.0
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycocotools==2.0
pycparser==2.21
Pygments==2.16.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-gan-metrics==0.5.3
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.12.25
requests==2.28.1
rich==13.7.0
safetensors==0.4.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.11.2
seaborn==0.13.0
sentry-sdk==1.31.0
setproctitle==1.3.2
shapely==2.0.2
six==1.16.0
smmap==5.0.1
soupsieve==2.5
stack-data==0.6.2
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.9.18
timm==0.9.12
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
typing_extensions==4.4.0
tzdata==2023.3
ultralytics==8.0.218
urllib3==1.26.13
visdom==0.2.4
wandb==0.15.11
wcwidth==0.2.6
websocket-client==1.6.3
Werkzeug==3.0.2
wrapt==1.16.0
xtcocotools==1.14.3
yacs==0.1.8
yapf==0.40.2
zipp==3.17.0absl-py==2.1.0
addict==2.4.0
appdirs==1.4.4
asttokens==2.4.0
astunparse==1.6.3
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
chumpy==0.70
click==8.1.7
colorama==0.4.6
colour==0.1.5
comm==0.2.0
contourpy==1.1.1
crowdposetools==2.0
cycler==0.11.0
Cython==3.0.2
debugpy==1.8.0
decorator==5.1.1
docker-pycreds==0.4.0
dominate==2.8.0
easydict==1.7
einops==0.7.0
exceptiongroup==1.1.3
executing==1.2.0
ffmpeg==1.4
filelock==3.9.0
flatbuffers==24.3.25
fonttools==4.42.1
fsspec==2023.12.2
gast==0.5.4
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.37
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
huggingface-hub==0.20.1
idna==3.4
imageio==2.31.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.26.0
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json-tricks==3.17.3
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.0
jupyter_core==5.5.0
keras==3.1.1
kiwisolver==1.4.5
lazy_loader==0.3
libclang==18.1.1
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mat4py==0.6.0
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
mmdet==3.2.0
mmengine==0.10.2
mmpose==1.3.0
mmpretrain==1.2.0
model-index==0.1.11
modelindex==0.0.2
mpmath==1.3.0
munch==4.0.0
munkres==1.1.4
namex==0.0.7
nest-asyncio==1.5.8
networkx==3.0
numpy==1.24.1
onnx==1.15.0
opencv-python==4.8.0.76
opt-einsum==3.3.0
optree==0.11.0
ordered-set==4.1.0
packaging==23.1
pandas==2.1.1
parso==0.8.3
pathtools==0.1.2
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==4.0.0
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycocotools==2.0
pycparser==2.21
Pygments==2.16.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-gan-metrics==0.5.3
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.12.25
requests==2.28.1
rich==13.7.0
safetensors==0.4.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.11.2
seaborn==0.13.0
sentry-sdk==1.31.0
setproctitle==1.3.2
shapely==2.0.2
six==1.16.0
smmap==5.0.1
soupsieve==2.5
stack-data==0.6.2
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.9.18
timm==0.9.12
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
typing_extensions==4.4.0
tzdata==2023.3
ultralytics==8.0.218
urllib3==1.26.13
visdom==0.2.4
wandb==0.15.11
wcwidth==0.2.6
websocket-client==1.6.3
Werkzeug==3.0.2
wrapt==1.16.0
xtcocotools==1.14.3
yacs==0.1.8
yapf==0.40.2
zipp==3.17.0
